fix(parquet): clean up byte-array page-sizing dead code + lint
Fixes the CI failures from #16 (cargo fmt + clippy unnecessary_cast in
arrow_writer_layout.rs) and applies that PR's follow-up recommendation:
remove the unreachable count_within_budget_accessor free function and its
`_ =>` accessor dispatch arm in ByteArrayEncoder.

The arm was dead code: every byte-array type ByteArrayEncoder is built for
has an explicit match arm, and a Dictionary(value = FixedSizeBinary) column
hits the Dictionary(_, _) arm (its values.data_type() is Dictionary). A bare
FixedSizeBinary column is routed to the generic column writer, never this
encoder. The arm is now an unreachable! with that rationale, and the
inaccurate "FixedSizeBinary falls through to the per-value walk" comment is
dropped.

+++ parquet/tests/arrow_writer_layout.rs
@@ -627,7 +627,8 @@ fn test_fixed_size_binary() {
         .flat_map(|i| vec![i as u8; value_size])
         .collect();
     let array =
-        Arc::new(FixedSizeBinaryArray::try_new(value_size as i32, values.into(), None).unwrap()) as _;
+        Arc::new(FixedSizeBinaryArray::try_new(value_size as i32, values.into(), None).unwrap())
+            as _;
     let batch = RecordBatch::try_from_iter([("col", array)]).unwrap();
 
     let props = WriterProperties::builder()
@@ -683,9 +684,9 @@ fn test_dictionary() {
     // column (one small dictionary page + one RLE_DICTIONARY data page).
     let num_rows = 2000;
     let dict_values = StringArray::from_iter_values(["alpha", "beta", "gamma", "delta"]);
-    let keys = Int32Array::from_iter_values((0..num_rows as i32).map(|i| i % 4));
-    let array = Arc::new(DictionaryArray::<Int32Type>::try_new(keys, Arc::new(dict_values)).unwrap())
-        as _;
+    let keys = Int32Array::from_iter_values((0..num_rows).map(|i| i % 4));
+    let array =
+        Arc::new(DictionaryArray::<Int32Type>::try_new(keys, Arc::new(dict_values)).unwrap()) as _;
     let batch = RecordBatch::try_from_iter([("col", array)]).unwrap();
 
     let props = WriterProperties::builder()
@@ -721,4 +722,3 @@ fn test_dictionary() {
         },
     });
 }
-
